Refactor, move useChangeFont to own file
